BillingCicleBackend/requirements.txt
certifi==2019.6.16
chardet==3.0.4
Django==2.2.3
djangorestframework==3.9.4
geographiclib==1.49
gunicorn==19.9.0
idna==2.8
psycopg2==2.8.3
pytz==2019.1
sqlparse==0.3.0
urllib3==1.25.3
whitenoise==4.1.2
